[core] Add dedicated 'Macro.h' header file.
