+++ core/include/jiminy/core/robot/BasicSensors.h
@@ -3,7 +3,7 @@
 
 #include "jiminy/core/robot/AbstractMotor.h"
 #include "jiminy/core/robot/AbstractSensor.h"
-#include "jiminy/core/Utilities.h"
+#include "jiminy/core/Macro.h"
 
 
 namespace jiminy
