+++ core/include/jiminy/core/telemetry/TelemetryData.h
@@ -11,6 +11,7 @@
 #include <unordered_map>
 
 #include "jiminy/core/telemetry/TelemetrySender.h"
+#include "jiminy/core/Macro.h"
 #include "jiminy/core/Types.h"
