+++ core/src/control/AbstractController.cc
@@ -3,7 +3,7 @@
 #include "pinocchio/algorithm/joint-configuration.hpp"
 
 #include "jiminy/core/robot/Robot.h"
-#include "jiminy/core/Utilities.h"
+#include "jiminy/core/Macro.h"
 #include "jiminy/core/Constants.h"
 
 #include "jiminy/core/control/AbstractController.h"
