+++ core/src/io/AbstractIODevice.cc
@@ -6,7 +6,7 @@
 
 #include <iostream>
 
-#include "jiminy/core/Utilities.h"
+#include "jiminy/core/Macro.h"
 #include "jiminy/core/io/AbstractIODevice.h"
