+++ core/src/robot/AbstractMotor.cc
@@ -1,5 +1,5 @@
 #include "jiminy/core/robot/Robot.h"
-#include "jiminy/core/Utilities.h"
+#include "jiminy/core/Macro.h"
 
 #include "jiminy/core/robot/AbstractMotor.h"
